Wind icon appears in possession when affected by Wind  (#3410)
