+++ src/config_magic.h
@@ -88,6 +88,7 @@ enum CreatureSpellAffectedFlags {
     /** For creature which are normally flying, this informs that its grounded due to spells or its condition. */
     CSAfF_Grounded     = 0x8000,
     CSAfF_Timebomb     = 0x10000,
+    CSAfF_Wind         = 0x20000,
 };
 
 enum PowerKinds {
