+++ src/thing_shots.h
@@ -110,6 +110,8 @@ TbBool detonate_shot(struct Thing *shotng, TbBool destroy);
 TbBool shot_is_boulder(const struct Thing *shotng);
 
 struct Thing *get_thing_collided_with_at_satisfying_filter(struct Thing *thing, struct Coord3d *pos, Thing_Collide_Func filter, HitTargetFlags a4, long a5);
+
+void affect_nearby_enemy_creatures_with_wind(struct Thing *thing);
 /******************************************************************************/
 #ifdef __cplusplus
 }
